% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{autoplot.tbl_cf}
\alias{autoplot.tbl_cf}
\title{Auto- and Cross- Covariance and -Correlation plots}
\usage{
\method{autoplot}{tbl_cf}(object, level = 95, ...)
}
\arguments{
\item{object}{A tbl_cf object (the result \code{\link[feasts:ACF]{feasts::ACF()}}, \code{\link[feasts:ACF]{feasts::PACF()}}, or \code{\link[feasts:ACF]{feasts::CCF()}}).}

\item{level}{The level of confidence for the blue dashed lines.}

\item{...}{Unused.}
}
\value{
A ggplot object showing the correlations.
}
\description{
Produces an appropriate plot for the result of  \code{\link[feasts:ACF]{feasts::ACF()}}, \code{\link[feasts:ACF]{feasts::PACF()}}, or \code{\link[feasts:ACF]{feasts::CCF()}}.
}
