% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-loop.R
\name{check_can_clip}
\alias{check_can_clip}
\title{Check that clipping grobs are supported}
\usage{
check_can_clip(context, check)
}
\arguments{
\item{context}{string giving the context (e.g. function, arguments, etc)
to use in message in case of failure}

\item{check}{if not \code{TRUE} no check is made}
}
\value{
\code{invisible(NULL)} if clipping is supported or raises an error if not.
}
\description{
Check that clipping grobs are supported
}
