% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-candlestick.R
\name{draw_key_candlestick}
\alias{draw_key_candlestick}
\title{Legend key for candlesticks}
\usage{
draw_key_candlestick(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grob.
}
\description{
A body spanning the middle of the key, with the high-low line running the
full height of it either side.
}
\keyword{internal}
