% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtime-package.R
\docType{package}
\name{ggtime-package}
\alias{ggtime}
\alias{ggtime-package}
\title{ggtime: Grammar of Graphics and Plot Helpers for Time Series Visualization}
\description{
Extends the capabilities of 'ggplot2' by providing grammatical elements and plot helpers designed for visualizing temporal patterns. The package implements a grammar of temporal graphics, which leverages calendar structures to highlight changes over time. The package also provides plot helper functions to quickly produce commonly used time series graphics, including time plots, season plots, and seasonal sub-series plots.
}
\author{
\strong{Maintainer}: ggtime authors \email{redacted@chi2026.acm.org}

}
\keyword{internal}
