% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistas-package.R
\docType{package}
\name{vistas-package}
\alias{vistas}
\alias{vistas-package}
\title{vistas: Visualisation of time series}
\description{
Visualisation of time series using ggplot2. Provides tools for graphically analysing time series, with exploration of trend and seasonality.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
}

}
\keyword{internal}
