% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-calendar.R
\name{coord_calendar}
\alias{coord_calendar}
\title{Calendar coordinates}
\usage{
coord_calendar(
  period = "week",
  time = "x",
  xlim = NULL,
  ylim = NULL,
  expand = TRUE,
  default = FALSE,
  clip = "on",
  coord = coord_cartesian()
)
}
\description{
Calendar coordinates
}
\examples{

library(ggplot2)
x <- 1:100
p <- tibble::tibble(
  time = Sys.time() + x*3600,
  sin = sin(x/pi)
) |>
  ggplot(aes(x = time, y = sin)) +
  geom_line()

p
p + coord_calendar(coord=coord_cartesian())
p + coord_calendar(coord=coord_polar())

}
