% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-calendar.R
\name{coord_calendar}
\alias{coord_calendar}
\title{Calendar coordinates}
\usage{
coord_calendar(
  loops = waiver(),
  time_loops = waiver(),
  time = "x",
  ljust = 0.5,
  xlim = NULL,
  ylim = NULL,
  expand = FALSE,
  default = FALSE,
  clip = "on",
  clip_loops = "on",
  coord = coord_cartesian()
)
}
\arguments{
\item{loops}{Loop the time scale around a calendrical granularity, one of:
\itemize{
\item \code{NULL} or \code{waiver()} for no looping (the default)
\item A \code{mixtime} vector giving time points at which the \code{time} axis should loop
\item A function that takes the limits as input and returns loop points as output
}}

\item{time_loops}{A duration giving the distance between temporal
loops like "2 weeks", or "10 years". If both \code{loops} and \code{time_loops} are
specified, \code{time_loops} wins.}

\item{time}{A string specifying which aesthetic contains the time variable that
should be looped over. Default is \code{"x"}.}

\item{ljust}{Loop justification, a number between 0 and 1
indicating where the lines between looped positions are drawn
(0 indicates left, 1 indicates right, 0.5 indicates center).}

\item{xlim, ylim}{Limits for the x and y axes. \code{NULL} means use the default limits.}

\item{expand}{Logical indicating whether to expand the coordinate limits.
Default is \code{FALSE}.}

\item{default}{Logical indicating whether this is the default coordinate system.
Default is \code{FALSE}.}

\item{clip}{Should drawing be clipped to the extent of the plot panel?
A setting of \code{"on"} (the default) means yes, and a setting of \code{"off"} means no.}

\item{clip_loops}{Should the drawing of each loop of the timescale be clipped to
the breaks defined by \code{time_loops} and \code{ljust}?
A setting of \code{"on"} (the default) means yes, and a setting of \code{"off"} means no.}

\item{coord}{The underlying coordinate system to use. Default is \code{coord_cartesian()}.}
}
\description{
Create a coordinate system that wraps time series data over specified periods.
}
