% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-calendar.R
\name{coord_calendar}
\alias{coord_calendar}
\title{Calendar coordinates}
\usage{
coord_calendar(
  period = "week",
  xlim = NULL,
  ylim = NULL,
  expand = TRUE,
  default = FALSE,
  clip = "on"
)
}
\description{
Calendar coordinates
}
