% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-loop.R
\name{coord_loop}
\alias{coord_loop}
\title{Looped coordinates}
\usage{
coord_loop(
  loop = waiver(),
  time_loop = waiver(),
  time = "x",
  xlim = NULL,
  ylim = NULL,
  expand = TRUE,
  default = FALSE,
  clip = "on",
  coord = coord_cartesian()
)
}
\description{
Looped coordinates
}
\examples{

library(ggplot2)
p <- as_tsibble(USAccDeaths) |>
  ggplot(aes(x = index, y = value)) +
  geom_line()

p
p + coord_loop(period = "year")

}
