% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-time.R
\name{scale_x_mixtime}
\alias{scale_x_mixtime}
\title{Position scales for mixtime data}
\usage{
scale_x_mixtime(
  name = waiver(),
  breaks = waiver(),
  time_breaks = waiver(),
  labels = waiver(),
  time_labels = waiver(),
  minor_breaks = waiver(),
  time_minor_breaks = waiver(),
  common_time = waiver(),
  warps = waiver(),
  time_warps = waiver(),
  limits = NULL,
  expand = waiver(),
  oob = scales::censor,
  guide = waiver(),
  position = "bottom",
  sec.axis = waiver()
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the breaks specified by \code{date_breaks}
\item A \code{Date}/\code{POSIXct} vector giving positions of breaks
\item A function that takes the limits as input and returns breaks as output
}}

\item{time_breaks}{A duration giving the distance between breaks like
"2 weeks", or "10 years". If both \code{breaks} and \code{time_breaks} are specified,
\code{time_breaks} wins.}

\item{labels}{One of the options below. Please note that when \code{labels} is a
vector, it is highly recommended to also set the \code{breaks} argument as a
vector to protect against unintended mismatches.
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{time_labels}{Uses strftime strings or similar to format the labels from
the time points.}

\item{minor_breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the breaks specified by \code{date_minor_breaks}
\item A \code{Date}/\code{POSIXct} vector giving positions of minor breaks
\item A function that takes the limits as input and returns minor breaks as
output
}}

\item{time_minor_breaks}{A duration giving the distance between minor breaks like
"2 weeks", or "10 years". If both \code{minor_breaks} and \code{time_minor_breaks} are
specified, \code{time_minor_breaks} wins.}

\item{common_time}{Acts like a vctrs \code{ptype} defining the common chronon to
use for mixed granularity. The default automatically selects it.}

\item{warps}{Warp the time scale to have a consistent length, one of:
\itemize{
\item \code{NULL} or \code{waiver()} for no warping (the default)
\item A \code{mixtime} vector giving positions of warping points
\item A function that takes the limits as input and returns warping points as
output
}}

\item{time_warps}{A duration giving the distance between temporal warping
like "2 weeks", or "10 years". If both \code{warps} and \code{time_warps} are
specified, \code{time_warps} wins.}

\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
Note that setting limits on positional scales will \strong{remove} data outside of the limits.
If the purpose is to zoom, use the limit argument in the coordinate system
(see \code{\link[ggplot2:coord_cartesian]{coord_cartesian()}}).
}}

\item{expand}{For position scales, a vector of range expansion constants used to add some
padding around the data to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[ggplot2:expansion]{expansion()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}

\item{oob}{One of:
\itemize{
\item Function that handles limits outside of the scale limits
(out of bounds). Also accepts rlang \link[rlang:as_function]{lambda}
function notation.
\item The default (\code{\link[scales:oob]{scales::censor()}}) replaces out of
bounds values with \code{NA}.
\item \code{\link[scales:oob]{scales::squish()}} for squishing out of bounds values into range.
\item \code{\link[scales:oob]{scales::squish_infinite()}} for squishing infinite values into range.
}}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{position}{For position scales, The position of the axis.
\code{left} or \code{right} for y axes, \code{top} or \code{bottom} for x axes.}

\item{sec.axis}{\code{\link[ggplot2:sec_axis]{sec_axis()}} is used to specify a secondary axis.}
}
\description{
These are the default scales for mixtime vectors, responsible for mapping
time points to aesthetics along with identifying break points and labels for
the axes and guides. To override the scales behaviour manually, use
\verb{scale_*_mixtime}.
}
\details{
Additional to the standard ggplot2 continuous scale options, these scales
provide time specific options including:

\strong{Granularity alignment}

Visualising mixed granularity time data introduces indeterminacy in the
mapping of less precise time points onto a common time scale. For example,
plotting monthly and daily data together raises the question of where to
place the monthly points relative to the daily points. By default, mixtime
uses center alignment, mapping the monthly points to the middle of the
month. This is controlled using the \code{time_align} argument, which accepts a
value between 0 (start alignment) and 1 (end alignment) and defaults to 0.5.

The common time scale that defines how all granularities are mapped is
automatically identified based on the input data. This is achieved by finding
the finest chronon that all time points can be represented in. For example,
if the data contains both monthly and daily time points, the common time scale
will be daily, with the monthly points aligned according to the \code{time_align}
argument. If multiple time zones are present, the common time zone will
default to UTC. The common time scale can be manually specified using the
\code{common_time} argument, which accepts a \code{mixtime::time_unit}.

\strong{Temporal warping}

Time scales can be warped to have a consistent length between specified time
points. This is useful when visually exploring cyclical patterns where each
cycle has varying length. By warping the time scale, the shape of each cycle
can be more easily compared. Temporal warping is controlled using the \code{warps}
argument, which accepts a \code{mixtime} vector defining the positions of the
warping points. Calendar-based warping points can be conveniently specified
using the \code{time_warps} argument, which accepts a duration like "1 month".
}
