% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-time.R
\name{scale_x_mixtime}
\alias{scale_x_mixtime}
\title{Position scales for mixtime data}
\usage{
scale_x_mixtime(
  name = waiver(),
  breaks = waiver(),
  time_breaks = waiver(),
  labels = waiver(),
  time_labels = waiver(),
  minor_breaks = waiver(),
  time_minor_breaks = waiver(),
  time_type = c("civil", "absolute"),
  wraps = waiver(),
  time_wraps = waiver(),
  warps = waiver(),
  time_warps = waiver(),
  limits = NULL,
  expand = waiver(),
  oob = scales::censor,
  guide = waiver(),
  position = "bottom",
  sec.axis = waiver()
)
}
\arguments{
\item{time_breaks}{A \if{html}{\out{<duration>}} giving the distance between breaks like
"2 weeks", or "10 years". If both \code{breaks} and \code{time_breaks} are specified,
\code{time_breaks} wins.}

\item{time_minor_breaks}{A \if{html}{\out{<duration>}} giving the distance between minor breaks like
"2 weeks", or "10 years". If both \code{minor_breaks} and \code{time_minor_breaks} are
specified, \code{time_minor_breaks} wins.}

\item{time_type}{The positioning of zoned time, "civil" uses the local/wall
time (the default) and "absolute" uses the continuous UTC time.}

\item{wraps}{Wrap the time scale around a calendrical granularity, one of:
\itemize{
\item \code{NULL} or \code{waiver()} for no wrapping (the default)
\item A \code{mixtime} vector giving positions of wrapping points
\item A function that takes the limits as input and returns wrapping points as
output
}}

\item{time_wraps}{A \if{html}{\out{<duration>}} giving the distance between temporal wrapping
like "2 weeks", or "10 years". If both \code{wraps} and \code{time_wraps} are
specified, \code{time_wraps} wins.}

\item{warps}{Warp the time scale to have a consistent length, one of:
\itemize{
\item \code{NULL} or \code{waiver()} for no warping (the default)
\item A \code{mixtime} vector giving positions of warping points
\item A function that takes the limits as input and returns warping points as
output
}}

\item{time_warps}{A \if{html}{\out{<duration>}} giving the distance between temporal warping
like "2 weeks", or "10 years". If both \code{warps} and \code{time_warps} are
specified, \code{time_warps} wins.}
}
\description{
These are the default scales for mixtime vectors, which are usually used for
a plot automatically. To override the scales behaviour manually, use
\verb{scale_*_mixtime}.
}
